Add `useFetchRepos` hook
Implementing DRY into `Popular` and isolating the Request Layer within the hook
